requirements.txt
streamlit
pydub
boto3
requests
streamlit-audiorecorder
python-dotenv
